CRUD - Category

+++ src/app/category/list/list-category.component.ts
@@ -1,43 +1,57 @@
 import { Component, OnInit } from '@angular/core';
-import { CategoryService } from '../category.service';
 import { Category } from '../category.model';
+import { CategoryService } from '../category.service';
 import { Router } from '@angular/router';
 
 @Component({
-  selector: 'app-list-category',
+  selector: 'app-category-list',
   templateUrl: './list-category.component.html',
 })
 export class ListCategoryComponent implements OnInit {
   categories: Category[] = [];
+  errorMessage: string | null = null;
 
-  constructor(
-    private categoryService: CategoryService,
-    private router: Router // Inject Router ở đây
-  ) {}
+  constructor(private categoryService: CategoryService, private router: Router) {}
 
   ngOnInit(): void {
     this.loadCategories();
   }
 
   loadCategories(): void {
-    this.categoryService.getAllCategories().subscribe({
-      next: (data) => this.categories = data,
-      error: (error) => console.error('Error loading categories:', error)
-    });
+    this.categoryService.getAllCategories().subscribe(
+      (categories) => {
+        this.categories = categories;
+        console.log(categories)
+        this.errorMessage = null;
+      },
+      (error) => {
+        console.error('Error loading products:', error);
+        this.errorMessage = 'Failed to load products: ' + error.message;
+      }
+    );
   }
-  
+
   editCategory(id: string): void {
-    this.router.navigate([`/categories/edit/${id}`]); // Sử dụng this.router để điều hướng
+    this.router.navigate(['/categories/edit', id]);
   }
 
   deleteCategory(id: string): void {
-    this.categoryService.deleteCategory(id).subscribe(
-      (success) => {
-        if (success) {
-          this.categories = this.categories.filter(cat => cat.id !== id);
+    const confirmDelete = confirm('Are you sure you want to delete this category?');
+    if (confirmDelete) {
+      this.categoryService.deleteCategory(id).subscribe(
+        () => {
+          this.loadCategories();
+          this.errorMessage = null;
+        },
+        (error) => {
+          console.error('Error deleting product:', error);
+          this.errorMessage = 'Failed to delete product: ' + error.message;
         }
-      },
-      (error) => console.error('Error deleting category:', error)
-    );
+      );
+    }
+  }
+
+  addNewCategory(): void {
+    this.router.navigate(['/categories/create']); // Chuyển hướng tới trang tạo sản phẩm
   }
 }
